% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_quanteda_app}
\alias{run_quanteda_app}
\title{Run Quanteda GUI App}
\usage{
run_quanteda_app(...)
}
\value{
A Shiny app object
}
\description{
This function launches a Shiny app that provides a graphical user interface for the Quanteda package.
}
\examples{
if(interactive()){
  run_quanteda_app()
}
}
